UI: Split up authentication endpoints and machine endpoints
