+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/home/page.tsx
@@ -1,17 +1,22 @@
 'use client';
-import { useGetAllMachinesQuery } from '@/redux/features/api/authentication/islandApi';
 import { useRouter } from 'next/navigation';
+import { useGetAllMachinesQuery } from '@/redux/features/api/machinesEndpoints';
+import { useEffect } from 'react';
 
 const HomePage = () => {
     const router = useRouter();
-    const { data, error, isLoading, isError, isSuccess } = useGetAllMachinesQuery();
+    const { data, error, isLoading, isError, isSuccess } =
+        useGetAllMachinesQuery();
+
+    useEffect(() => {
+        if (isError && error.status === 401) {
+            router.push('/signin'); // redirect to login page
+        }
+    }, [isError, error, router]);
 
     if (isSuccess) return <div>{MachineList(data)}</div>;
     if (isLoading) return <div>loading...</div>;
-    if (isError && error.status === 401) {
-        router.push('/signin'); // redirect to login page
-    }
-    else if (isError) {
+    if (isError) {
         return <div>Error: {error.data.response.errors}</div>;
     }
 };
